requirements.txt
asgiref==3.2.7
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httpie==2.1.0
httplib2==0.18.1
idna==2.9
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
